feat : TodoController와 DetailTodoController에 다른 Model, View를 연결
- TodoController와 DetailTodoController의 동작에 의해 외부 Model과 View도 변화
- 따라서, 외부 Model과 View를 연결하여 데이터를 처리해주는 과정을 만들었다.
- 이 과정에서 동작의 흐름이 복잡해지고 코드의 크기가 점차 커지는 것을 확인할 수 있었다.

+++ MVC_Pattern/Todo-list-app/src/controller/DetailTodoController.ts
@@ -1,14 +1,33 @@
 import DetailTodoModel from "../model/DetailTodoModel";
+import TodoCounterModel from "../model/TodoCounterModel";
+import TodoModel from "../model/TodoModel";
 import { TodoElement } from "../types/todo";
 import DetailTodoView from "../view/DetailTodoView";
+import TodoCounterView from "../view/TodoCounterView";
+import TodoView from "../view/TodoView";
 
 class DetailTodoController {
   private detailTodoModel: DetailTodoModel;
   private detailTodoView: DetailTodoView;
+  private todoModel: TodoModel;
+  private todoView: TodoView;
+  private todoCounterModel: TodoCounterModel;
+  private todoCounterView: TodoCounterView;
 
-  constructor(detailTodoModel: DetailTodoModel, detailTodoView: DetailTodoView) {
+  constructor(
+    detailTodoModel: DetailTodoModel,
+    detailTodoView: DetailTodoView,
+    todoModel: TodoModel,
+    todoView: TodoView,
+    todoCounterModel: TodoCounterModel,
+    todoCounterView: TodoCounterView
+  ) {
     this.detailTodoModel = detailTodoModel;
     this.detailTodoView = detailTodoView;
+    this.todoModel = todoModel;
+    this.todoView = todoView;
+    this.todoCounterModel = todoCounterModel;
+    this.todoCounterView = todoCounterView;
     this.render();
     this.detailTodoView.bindTodoDeleteButton(this.handleDeleteTodo.bind(this));
     this.detailTodoView.bindTodoStateToggleButton(this.handleToggleTodoState.bind(this));
@@ -18,14 +37,37 @@ class DetailTodoController {
     this.detailTodoView.displayDetailTodo(this.detailTodoModel.getDetailedTodo());
   }
 
+  renderTodo() {
+    this.todoView.displayTodos(this.todoModel.getAllTodos());
+  }
+
+  renderCounter() {
+    this.todoCounterView.displayTodoCounter(this.todoCounterModel.getCounterData());
+  }
+
   handleDeleteTodo() {
+    const detailTodo = this.detailTodoModel.getDetailedTodo();
+    if (!detailTodo) throw Error("invalid delete event");
+
     this.detailTodoModel.deleteDetailedTodo();
     this.render();
+
+    this.todoModel.deleteTodo(detailTodo.id);
+    this.renderTodo();
+
+    this.todoCounterModel.countDeletedTodo(detailTodo.state);
+    this.renderCounter();
   }
 
   handleToggleTodoState() {
+    const detailTodo = this.detailTodoModel.getDetailedTodo();
+    if (!detailTodo) throw Error("invalid delete event");
+
     this.detailTodoModel.toggleDetailTodoState();
     this.render();
+
+    this.todoCounterModel.countToggledTodoState(detailTodo.state);
+    this.renderCounter();
   }
 
   handleSetDetailTodo(todo: TodoElement) {
